requirements.txt
boto3==1.13.14
botocore==1.20.0
cryptography==3.3.2
cuckoofilter==0.0.0.2
filesplit==4.0.1
pycryptodome==3.17
pyfiglet==0.8.post1
rsa==4.0
